WEB3-46 - native contracts table renamed to reserved_addresses
