+++ apps/block_scout_web/lib/block_scout_web/templates/smart_contract/_functions.html.eex
@@ -1,7 +1,7 @@
 <% minimal_proxy_template = if assigns[:custom_abi], do: nil, else: Chain.get_minimal_proxy_template(@address.hash) %>
 <% metadata_for_verification = if assigns[:custom_abi], do: nil, else: minimal_proxy_template || Chain.get_address_verified_twin_contract(@address.hash).verified_contract %>
 <% smart_contract_verified = if assigns[:custom_abi], do: false, else: BlockScoutWeb.AddressView.smart_contract_verified?(@address) %>
-<% smart_contract_native = BlockScoutWeb.NativeContractUtils.smart_contract_native?(@address) %>
+<% smart_contract_native = BlockScoutWeb.ReservedAddressesUtils.smart_contract_native?(@address) %>
 <%= unless smart_contract_verified do %>
   <%= if metadata_for_verification do %>
     <%= if minimal_proxy_template do %>
